Do not check for duplicate entry in a group
The idea was to extend this to check for duplicates in all
the vocabulary groups using all_words.json. But, now this looks
like an unnecessary task. The user should be given the freedom to
add words to any group, any number of times.

This also removes some pre processing overhead.